add more stuff to data and connect it to the scripts
makes things more neat
