--- dev/Unity/Assets/titleNavScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 121019e74eb39fc4ba1e8995b77b85f4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
