+++ dev/Unity/UserSettings/EditorUserSettings.asset
@@ -9,10 +9,10 @@ EditorUserSettings:
       value: 5700525f5c570d5a085b54701176594417164e2e75707f69792b4863b0b9676b
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 5a07575657070a080f565e27497b5944154f4f7a287e77667b281966e4b13168
+      value: 515301570606080f550b097a49725c4446164128797e22352b284566b3b0653e
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 515301570606080f550b097a49725c4446164128797e22352b284566b3b0653e
+      value: 5a07575657070a080f565e27497b5944154f4f7a287e77667b281966e4b13168
       flags: 0
     RecentlyUsedSceneGuid-3:
       value: 5605045054540a0d0b0f5b7a13700a444e4e497f7a7c7f317b7d4936b5b2363d
